requirements.txt
# Core Data Processing
pandas==2.1.4
numpy==1.24.3
requests==2.31.0
yfinance>=0.2.0

# Configuration Management
python-dotenv==1.0.0

# Visualization & Dashboard
streamlit==1.29.0
plotly==5.17.0
tabulate==0.9.0

# Development Tools
pytest==7.4.3
pytest-cov==4.1.0